pip-requirements.txt
absl-py==0.9.0
adal==1.2.2
alembic==1.4.1
amqp==2.5.2
aniso8601==8.0.0
appdirs==1.4.3
astor==0.8.1
async-generator==1.10
attrs==19.3.0
backcall==0.1.0
behave==1.2.6
billiard==3.5.0.5
bleach==3.1.1
braintree==3.59.0
cachetools==4.0.0
callable-pip==1.0.0
celery==4.2.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cron-descriptor==1.2.24
croniter==0.3.31
cryptography==2.8
cycler==0.10.0
dash==1.9.1
dash-core-components==1.8.1
dash-daq==0.4.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
dashserve==0.1.5
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
dj-database-url==0.4.2
Django==1.8.19
django-allauth==0.34.0
django-autoslug==1.9.6
django-cities-light==3.4.0
django-classy-tags==1.0.0
django-constance==2.0.0
django-debug-toolbar==1.9.1
django-extensions==1.7.7
django-mailgun==0.9.1
django-nose==1.4.5
django-organizations==1.0.0
django-payments==0.12.0
django-permissions==1.2.2
django-picklefield==2.0
django-post-office==3.0.4
django-sekizai==0.10.0
django-tastypie==0.12.2
docopt==0.6.2
docutils==0.15
entrypoints==0.3
escapism==1.0.0
Flask==1.1.1
Flask-Compress==1.4.0
flask-restplus==0.13.0
future==0.18.2
gast==0.2.2
google-auth==1.11.2
google-pasta==0.1.8
grpcio==1.27.2
h5py==2.10.0
httplib2==0.17.0
idna==2.7
imageio==2.6.1
importlib-metadata==1.5.0
invoke==1.4.1
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonfield==2.0.2
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
jupyterhub==0.9.4
jupyterhub-kubespawner==0.12.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kombu==4.6.8
kubernetes==8.0.0
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mock==3.0.5
mongoengine==0.18.2
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==5.7.6
numexpr==2.7.1
numpy==1.17.5
oauthlib==3.1.0
opt-einsum==3.2.0
ordered-set==3.1.1
pamela==1.0.0
pandas==1.0.1
pandocfilters==1.4.2
parse==1.15.0
parse-type==0.5.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.5.3
progressbar2==3.50.0
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pymongo==3.10.1
pyparsing==2.4.6
pyrabbit==1.1.0
pyrsistent==0.15.7
python-dateutil==2.8.1
python-editor==1.0.4
python-mimeparse==1.6.0
python-oauth2==1.1.1
python-utils==2.4.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
requests==2.20.0
requests-oauthlib==1.0.0
retrying==1.3.3
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
seaborn==0.9.0
selenium==3.141.0
Send2Trash==1.5.0
six==1.11.0
SQLAlchemy==1.3.14
sqlparse==0.3.1
stripe==2.43.0
suds-jurko==0.6
tables==3.6.1
tee==0.0.3
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.24.3
vine==1.3.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.16.1
whitenoise==3.3.1
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
wheel
setuptools
pyrabbit2==1.0.7
